+++ angular.json
@@ -3,7 +3,40 @@
     "version": 1,
     "newProjectRoot": "projects",
     "projects": {
-        "cookies-pref": {
+        "ngx-nc-cookies-pref": {
+            "projectType": "library",
+            "root": "projects/ngx-nc-cookies-pref",
+            "sourceRoot": "projects/ngx-nc-cookies-pref/src",
+            "prefix": "lib",
+            "architect": {
+                "build": {
+                    "builder": "@angular-devkit/build-angular:ng-packagr",
+                    "options": {
+                        "project": "projects/ngx-nc-cookies-pref/ng-package.json"
+                    },
+                    "configurations": {
+                        "production": {
+                            "tsConfig": "projects/ngx-nc-cookies-pref/tsconfig.lib.prod.json"
+                        },
+                        "development": {
+                            "tsConfig": "projects/ngx-nc-cookies-pref/tsconfig.lib.json"
+                        }
+                    },
+                    "defaultConfiguration": "production"
+                },
+                "test": {
+                    "builder": "@angular-devkit/build-angular:karma",
+                    "options": {
+                        "tsConfig": "projects/ngx-nc-cookies-pref/tsconfig.spec.json",
+                        "polyfills": [
+                            "zone.js",
+                            "zone.js/testing"
+                        ]
+                    }
+                }
+            }
+        },
+        "playground": {
             "projectType": "application",
             "schematics": {
                 "@schematics/angular:component": {
@@ -11,26 +44,26 @@
                 }
             },
             "root": "",
-            "sourceRoot": "src",
+            "sourceRoot": "projects/playground",
             "prefix": "app",
             "architect": {
                 "build": {
                     "builder": "@angular-devkit/build-angular:browser",
                     "options": {
-                        "outputPath": "dist/cookies-pref",
-                        "index": "src/index.html",
-                        "main": "src/main.ts",
+                        "outputPath": "dist/playground",
+                        "index": "projects/playground/src/index.html",
+                        "main": "projects/playground/src/main.ts",
                         "polyfills": [
                             "zone.js"
                         ],
                         "tsConfig": "tsconfig.app.json",
                         "inlineStyleLanguage": "scss",
                         "assets": [
-                            "src/favicon.ico",
-                            "src/assets"
+                            "projects/playground/src/favicon.ico",
+                            "projects/playground/src/assets"
                         ],
                         "styles": [
-                            "src/styles.scss",
+                            "projects/playground/src/styles.scss",
                             "node_modules/primeng/resources/themes/lara-light-blue/theme.css",
                             "node_modules/primeng/resources/primeng.min.css"
                         ],
@@ -67,10 +100,10 @@
                     "builder": "@angular-devkit/build-angular:dev-server",
                     "configurations": {
                         "production": {
-                          "buildTarget": "cookies-pref:build:production"
+                          "buildTarget": "playground:build:production"
                         },
                         "development": {
-                          "buildTarget": "cookies-pref:build:development"
+                          "buildTarget": "playground:build:development"
                         }
                     },
                     "defaultConfiguration": "development"
@@ -78,7 +111,7 @@
                 "extract-i18n": {
                     "builder": "@angular-devkit/build-angular:extract-i18n",
                     "options": {
-                      "buildTarget": "cookies-pref:build"
+                      "buildTarget": "playground:build"
                     }
                 },
                 "test": {
@@ -91,46 +124,13 @@
                         "tsConfig": "tsconfig.spec.json",
                         "inlineStyleLanguage": "scss",
                         "assets": [
-                            "src/favicon.ico",
-                            "src/assets"
+                            "projects/playground/src/favicon.ico",
+                            "projects/playground/src/assets"
                         ],
                         "styles": [
-                            "src/styles.scss"
+                            "projects/playground/src/styles.scss"
                         ],
                         "scripts": []
-                    }
-                }
-            }
-        },
-        "ngx-nc-cookies-pref": {
-            "projectType": "library",
-            "root": "projects/ngx-nc-cookies-pref",
-            "sourceRoot": "projects/ngx-nc-cookies-pref/src",
-            "prefix": "lib",
-            "architect": {
-                "build": {
-                    "builder": "@angular-devkit/build-angular:ng-packagr",
-                    "options": {
-                        "project": "projects/ngx-nc-cookies-pref/ng-package.json"
-                    },
-                    "configurations": {
-                        "production": {
-                            "tsConfig": "projects/ngx-nc-cookies-pref/tsconfig.lib.prod.json"
-                        },
-                        "development": {
-                            "tsConfig": "projects/ngx-nc-cookies-pref/tsconfig.lib.json"
-                        }
-                    },
-                    "defaultConfiguration": "production"
-                },
-                "test": {
-                    "builder": "@angular-devkit/build-angular:karma",
-                    "options": {
-                        "tsConfig": "projects/ngx-nc-cookies-pref/tsconfig.spec.json",
-                        "polyfills": [
-                            "zone.js",
-                            "zone.js/testing"
-                        ]
                     }
                 }
             }
