+++ projects/ngx-nc-cookies-pref/package.json
@@ -2,7 +2,7 @@
     "name": "ngx-cookies-preference",
     "author": "Nicola Cocozza",
     "description": "In a project angular set a cookies preferences.",
-    "version": "3.0.0",
+    "version": "3.0.1",
     "private": false,
     "keywords": [
         "cookies",
