+++ projects/playground/src/app/app.module.ts
@@ -2,9 +2,11 @@ import {NgModule} from '@angular/core';
 import {BrowserModule} from '@angular/platform-browser';
 
 import {AppComponent} from './app.component';
-import {PopupCookiesComponent} from '../../src/app/cookies-pref/components/popup-cookies/popup-cookies.component';
 import {TranslateModule} from '@ngx-translate/core';
 import {NgbModule} from '@ng-bootstrap/ng-bootstrap';
+import {
+    PopupCookiesComponent
+} from '../../../ngx-nc-cookies-pref/src/lib/cookies-pref/components/popup-cookies/popup-cookies.component';
 
 @NgModule({
     declarations: [
