+++ projects/playground/src/styles.scss
@@ -1,6 +1,6 @@
-@import "bootstrap/dist/css/bootstrap.css";
-@import "primeng/resources/themes/lara-light-blue/theme.css";
-@import "primeng/resources/primeng.css";
+@import "../../../node_modules/bootstrap/dist/css/bootstrap.css";
+@import "../../../node_modules/primeng/resources/themes/lara-light-blue/theme.css";
+@import "../../../node_modules/primeng/resources/primeng.css";
 
 /* You can add global styles to this file, and also import other style files */
 :host {
