--- src/app/cookies-pref/models/config-popup.model.ts
@@ -1,54 +1,0 @@
-export class ConfigPopup {
-
-    public colorPrincipalText?: string;
-    public textTitleBanner?: string;
-    public textDescriptionBanner?: string;
-    public textDescriptionBannerCut?: string;
-    public classBtn?: string;
-    public textBtnSetting?: string;
-    public textBtnAcceptEssential?: string;
-    public textBtnAcceptAll?: string;
-    public cookieKey?: string;
-    public textTitleModal?: string;
-    public textDescriptionModal?: string;
-    public textTitleTechnical?: string;
-    public textDescriptionTechnical?: string;
-    public textTitleFunctional?: string;
-    public textDescriptionFunctional?: string;
-    public textTitleMarketing?: string;
-    public textDescriptionMarketing?: string;
-    public textBtnAcceptSelected?: string;
-    public textReadMore?: string;
-    public sizeModal?: string;
-    public sizeTextSpaceModal?: string;
-    public classBtnContainer?: string;
-
-    constructor(colorPrincipalText?: string, textTitleBanner?: string, textDescriptionBanner?: string, textDescriptionBannerCut?: string,
-                classBtn?: string, textBtnSetting?: string, textBtnAcceptEssential?: string, textBtnAcceptAll?: string, cookieKey?: string,
-                textTitleModal?: string, textDescriptionModal?: string, textTitleTechnical?: string, textDescriptionTechnical?: string,
-                textTitleFunctional?: string, textDescriptionFunctional?: string, textTitleMarketing?: string, textDescriptionMarketing?: string,
-                textBtnAcceptSelected?: string, textReadMore?: string, sizeModal?: string, sizeTextSpaceModal?: string, classBtnContainer?: string) {
-        this.colorPrincipalText = colorPrincipalText;
-        this.textTitleBanner = textTitleBanner;
-        this.textDescriptionBanner = textDescriptionBanner;
-        this.textDescriptionBannerCut = textDescriptionBannerCut;
-        this.classBtn = classBtn;
-        this.textBtnSetting = textBtnSetting;
-        this.textBtnAcceptEssential = textBtnAcceptEssential;
-        this.textBtnAcceptAll = textBtnAcceptAll;
-        this.cookieKey = cookieKey;
-        this.textTitleModal = textTitleModal;
-        this.textDescriptionModal = textDescriptionModal;
-        this.textTitleTechnical = textTitleTechnical;
-        this.textDescriptionTechnical = textDescriptionTechnical;
-        this.textTitleFunctional = textTitleFunctional;
-        this.textDescriptionFunctional = textDescriptionFunctional;
-        this.textTitleMarketing = textTitleMarketing;
-        this.textDescriptionMarketing = textDescriptionMarketing;
-        this.textBtnAcceptSelected = textBtnAcceptSelected;
-        this.textReadMore = textReadMore;
-        this.sizeModal = sizeModal;
-        this.sizeTextSpaceModal = sizeTextSpaceModal;
-        this.classBtnContainer = classBtnContainer;
-    }
-}
