--- src/app/cookies-pref/models/cookies-pref.model.ts
@@ -1,11 +1,0 @@
-export class CookiesPref {
-    public technical: boolean;
-    public functional: boolean;
-    public marketing: boolean;
-
-    constructor(technical: boolean, functional: boolean, marketing: boolean) {
-        this.technical = technical;
-        this.functional = functional;
-        this.marketing = marketing;
-    }
-}
