+++ tsconfig.spec.json
@@ -8,7 +8,7 @@
     ]
   },
   "include": [
-    "src/**/*.spec.ts",
-    "src/**/*.d.ts"
+    "projects/playground/src/**/*.spec.ts",
+    "projects/playground/src/**/*.d.ts"
   ]
 }
